Avoid warnings during compilation.
[git-p4: depot-paths = "//depot/qt/main/": change = 72211]

+++ src/tools/qfile_unix.cpp
@@ -433,7 +433,7 @@ QIODevice::Offset QFile::size() const
 	::stat( QFile::encodeName(fn), &st );
     }
 #if defined(QT_LARGEFILE_SUPPORT) && !defined(QT_ABI_64BITOFFSET)
-    return st.st_size > UINT_MAX ? UINT_MAX : (QIODevice::Offset)st.st_size;
+    return (uint)st.st_size > UINT_MAX ? UINT_MAX : (QIODevice::Offset)st.st_size;
 #else
     return st.st_size;
 #endif
